requirements.txt
flask>=2.0,<3.0
flask-smorest>=0.42
python-dotenv
sqlalchemy>=2.0
flask-sqlalchemy>=3.1
flask-jwt-extended
passlib
flask-migrate>=4.0
gunicorn
psycopg2
requests
rq
redis